libs/arithmetic_python_client/requirements.txt
# Required packages
grpcio==1.47.0
grpcio-tools==1.47.0
protobuf==3.19.4

# Pytests
docker==6.0.0
